refactor: switch package locations
