+++ src/main/kotlin/me/tech/mcchestui/GUI.kt
@@ -7,6 +7,12 @@
 
 package me.tech.mcchestui
 
+import me.tech.mcchestui.listeners.GUIListener
+import me.tech.mcchestui.utils.GUICloseEvent
+import me.tech.mcchestui.utils.GUIDragItemEvent
+import me.tech.mcchestui.utils.GUIItemPickupEvent
+import me.tech.mcchestui.utils.GUIItemPlaceEvent
+import me.tech.mcchestui.utils.GUISlotClickEvent
 import net.kyori.adventure.text.Component
 import org.bukkit.entity.Player
 import org.bukkit.event.*
